Update Board.java
The first version of the game was built primarily for young children who could not read. This version accommodates readers as well. Consequently, in addition to the colored cards it already has, the game includes a second kind of card. These cards are not colored. Instead, the cards feature the name of the color. i.e. "red", "yellow", "green", "blue", "magenta", or "orange". A text card only matches another text card with the same color name, and a color card only matches another color card that features the same color. All other comparisons result in a mismatch.
When the game is started, the user specifies (in some convenient way) how many pairs of cards they want to use for that game. For example, the user may want 10 pairs of cards, in which case your game features 20 cards. The cards should be laid out in an arrangement as close to square as possible. For example, 20 cards would be laid out in a 5x4 pattern.
Each game includes a random mix of both kinds of cards.